requirements.txt
Django==4.2.14
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
Faker==26.0.0